refactor eos.py. move from io.abinit to analysis
