+++ pymatgen/io/abinit/__init__.py
@@ -2,9 +2,9 @@
 # Copyright (c) Pymatgen Development Team.
 # Distributed under the terms of the MIT License.
 
-from .eos import *
-from .pseudos import *
+from .calculations import *
 from .netcdf import *
+from .pseudos import *
 from .tasks import *
 from .works import *
-from .calculations import *
+
